+++ config/ember-try.js
@@ -1,23 +1,24 @@
 'use strict';
 
 const getChannelURL = require('ember-source-channel-url');
+const { embroiderSafe, embroiderOptimized } = require('@embroider/test-setup');
 
 module.exports = async function () {
   return {
     scenarios: [
       {
-        name: 'ember-lts-3.16',
+        name: 'ember-lts-3.24',
         npm: {
           devDependencies: {
-            'ember-source': '~3.16.0'
+            'ember-source': '~3.24.3'
           }
         }
       },
       {
-        name: 'ember-lts-3.20',
+        name: 'ember-lts-3.28',
         npm: {
           devDependencies: {
-            'ember-source': '~3.20.5'
+            'ember-source': '~3.28.0'
           }
         }
       },
@@ -49,12 +50,13 @@ module.exports = async function () {
         name: 'ember-default-with-jquery',
         env: {
           EMBER_OPTIONAL_FEATURES: JSON.stringify({
-            'jquery-integration': true
+            'jquery-integration': false
           })
         },
         npm: {
           devDependencies: {
-            '@ember/jquery': '^1.1.0'
+            '@ember/jquery': '^1.1.0',
+            '@ember/optional-features': '^2.0.0'
           }
         }
       },
@@ -68,11 +70,16 @@ module.exports = async function () {
           })
         },
         npm: {
+          devDependencies: {
+            'ember-source': '~3.28.0'
+          },
           ember: {
             edition: 'classic'
           }
         }
-      }
+      },
+      embroiderSafe(),
+      embroiderOptimized()
     ]
   };
 };