+++ tests/acceptance/r-test.js
@@ -1,5 +1,5 @@
 import { module, test } from 'qunit';
-import { visit, currentURL, typeIn, find } from '@ember/test-helpers';
+import { currentURL, typeIn, visit } from '@ember/test-helpers';
 import { setupApplicationTest } from 'ember-qunit';
 
 module('Acceptance | r', function (hooks) {
@@ -7,8 +7,8 @@ module('Acceptance | r', function (hooks) {
 
   test('can visit /r', async function (assert) {
     await visit('/r');
-    assert.equal(currentURL(), '/r');
+    assert.strictEqual(currentURL(), '/r');
     await typeIn('#dasherize', 'new value');
-    assert.equal(find('#result').textContent, 'new-value');
+    assert.dom('#result').hasText('new-value');
   });
 });